# docs/swagger.yaml
definitions:
...
  controllers.PostCreateVoucherRequest:
    properties:
      code:
        type: string
      description:
        type: string
      max:
        type: number
      name:
        type: string
      percentage:
        type: number
    required:
    - code
    - description
    - max
    - name
    - percentage
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a API for Online Shopping Cart API.
  title: Online Shopping Cart API
  version: "1.0"
paths:
...
  /voucher:
    post:
      description: add voucher for admin user, need credential ADMIN user only
      parameters:
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluQGV4YW1wbGUuY29tIiwiZW1haWwiOiJhZG1pbkBleGFtcGxlLmNvbSIsInR5cGUiOiJBRE1JTiIsImV4cCI6MTcyNTU5ODc3OX0.JamULnKlo3q38ZgIhfBOUI8U2WEv4nNfaLYvodtIx0c
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PostCreateVoucherRequest'
      produces:
      - application/json
      responses: {}
      summary: add voucher for admin user
      tags:
      - Voucher
  /voucher/{id}:
    delete:
      description: delete voucher for admin user, need credential ADMIN user only
      parameters:
      - description: delete by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluQGV4YW1wbGUuY29tIiwiZW1haWwiOiJhZG1pbkBleGFtcGxlLmNvbSIsInR5cGUiOiJBRE1JTiIsImV4cCI6MTcyNTU5ODc3OX0.JamULnKlo3q38ZgIhfBOUI8U2WEv4nNfaLYvodtIx0c
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: delete voucher for admin user
      tags:
      - Voucher
    get:
      description: get one voucher detail, need credentials
      parameters:
      - description: get detail by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: get one voucher detail
      tags:
      - Voucher
    put:
      description: edit voucher for admin user, need credential ADMIN user only
      parameters:
      - description: edit by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluQGV4YW1wbGUuY29tIiwiZW1haWwiOiJhZG1pbkBleGFtcGxlLmNvbSIsInR5cGUiOiJBRE1JTiIsImV4cCI6MTcyNTU5ODc3OX0.JamULnKlo3q38ZgIhfBOUI8U2WEv4nNfaLYvodtIx0c
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PostCreateVoucherRequest'
      produces:
      - application/json
      responses: {}
      summary: edit voucher for admin user
      tags:
      - Voucher
  /voucher/all:
    get:
      description: get all voucher need credentials
      parameters:
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: code voucher
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses: {}
      summary: get all voucher
      tags:
      - Voucher
swagger: "2.0"